Move all style files to folders in each directory
