+++ src/components/calculator/PointsDisplay.tsx
@@ -12,7 +12,7 @@ import {
   showPoints,
 } from 'components/calculator/util/displayUtil'
 import { ONE_MILLION } from 'data/calculate'
-import './PointsDisplay.scss'
+import './styles/PointsDisplay.scss'
 
 export interface RiskLevel {
   style: string
