+++ src/components/calculator/styles/ControlLabel.scss
@@ -1,5 +1,5 @@
-@import '../../styles/_variables.scss';
-@import '../../styles/breakpoints.scss';
+@import '../../../styles/_variables.scss';
+@import '../../../styles/breakpoints.scss';
 @import "~rfs/scss";
 
 // Tooltip button/badge styles
