+++ src/components/calculator/styles/PointsDisplay.scss
@@ -1,5 +1,5 @@
-@import '../../styles/_variables.scss';
-@import '../../styles/breakpoints.scss';
+@import '../../../styles/_variables.scss';
+@import '../../../styles/breakpoints.scss';
 @import "~rfs/scss";
 
 $sticky-z-index: 100;
