+++ src/components/calculator/styles/PrevalenceControls.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 #prevelance-details {
   .prevelance-result.card {
