+++ src/components/styles/Expandable.scss
