+++ src/pages/styles/Paper.scss
@@ -1,5 +1,5 @@
-@import '../styles/_variables.scss';
-@import '../styles/breakpoints.scss';
+@import '../../styles/_variables.scss';
+@import '../../styles/breakpoints.scss';
 
 .paperPage {
   font-size: 18px;
